refactor(auth): migrate to cookie-based ModeratorProtectedRoute and remove legacy ProtectedRoute
